dissect-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.2.0
pdfplumber>=0.10.0
python-docx>=1.1.0
sse-starlette>=2.0.0
